requirements.txt
img2dataset
pandas
sqlalchemy
tqdm
pyyaml 